requirements.txt
RPi.GPIO
pyyaml

